Add search for hosts and networks (control with button)
Less full DOM redraw, more careful additions

User roles are enforced, using original PieNg role descriptions

Ping new host creation is now just default, and we use TCP SynAck or
RST response as a ping reply, instead of ICMP echo reply, so we don't
have to call an external tool. This is just my preference, as we
usually allocate hosts before they are up, so we want to warn people
that this is happening.

The health check and the "new information redisplay" check are now
merged.

Various other UI fixes and improvements